requirements.txt
langchain
langchain-community
faiss-cpu
transformers
sentence-transformers
gpt4all
pandas
tqdm
accelerate 
bitsandbytes